Convert TraceProducer into an interface
