+++ src/main/java/com/braintreepayments/apollo_tracing_uploader/impl/ScheduledBatchingTraceProducer.java
@@ -11,8 +11,8 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.stream.IntStream;
 
+import com.braintreepayments.apollo_tracing_uploader.FullTracesReportBuilder;
 import com.braintreepayments.apollo_tracing_uploader.TraceProducer;
-import com.braintreepayments.apollo_tracing_uploader.Uploader;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,8 +26,11 @@ import mdg.engine.proto.Reports;
  * <p>
  * {@link ScheduledBatchingTraceProducer#shutdown} should be called at application shutdown to prevent dropped metrics.
  */
-public class ScheduledBatchingTraceProducer extends TraceProducer {
+public class ScheduledBatchingTraceProducer implements TraceProducer {
   private final Logger logger = LoggerFactory.getLogger(ScheduledBatchingTraceProducer.class);
+  private final FullTracesReportBuilder reportBuilder = new FullTracesReportBuilder();
+  private final Consumer<Reports.ReportHeader.Builder> customizeReportHeader;
+  private final Uploader uploader;
   private final int threadPoolSize;
   private final BlockingQueue<Reports.Trace> queue;
   private final ScheduledExecutorService executor;
@@ -41,8 +44,8 @@ public class ScheduledBatchingTraceProducer extends TraceProducer {
                                         int threadPoolSize,
                                         Duration batchingWindow,
                                         BlockingQueue<Reports.Trace> queue) {
-    super(customizeReportHeader, uploader);
-
+    this.customizeReportHeader = customizeReportHeader;
+    this.uploader = uploader;
     this.threadPoolSize = threadPoolSize;
     this.queue = queue;
     this.executor = Executors.newScheduledThreadPool(threadPoolSize);
