+++ src/main/java/com/braintreepayments/apollo_tracing_uploader/impl/Uploader.java
@@ -1,0 +1,8 @@
+package com.braintreepayments.apollo_tracing_uploader.impl;
+
+import mdg.engine.proto.Reports;
+
+@FunctionalInterface
+public interface Uploader {
+  void upload(Reports.FullTracesReport report);
+}
